+++ transports/changelog.md
@@ -7,3 +7,6 @@
 - fix: azure openai sdk requests handling in openai integration
 - feat: support multiple credential types in vertex auth credentials
 - fix: semantic caching plugin initialization when configured using UI
+- fix: added replicate and huggingface model allowlist on list models response
+- fix: added support for unfiltered list models response when updating keys' allowed models field
+- fix: backfill allowed models that were not in the list models response